Update dataset with the survey ID for each point (#3)
In case we want to investigate or extract individual surveys, we need to
preserve this information somehow. While I'm at it, I'll also update the sample
figures and generate a CSV version as well for people who don't want to deal
with netCDF and don't need the metadata. One caveat is that one of the survey
files was missing from the server so this compilation is a bit smaller (only a
few 100 points).